requirements.txt
asgiref==3.2.7
astroid==2.3.3
basic-api==0.1.1
certifi==2019.11.28
chardet==3.0.4
contextlib2==0.5.5
Django==3.0.5
django-filter==2.2.0
djangorestframework==3.11.0
docopt==0.6.2
gunicorn==20.0.4
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.2.1
mccabe==0.6.1
psycopg2-binary==2.8.4
pylint==2.4.4
pylint-django==2.0.15
pylint-plugin-utils==0.6
python-dateutil==2.8.1
pythonanywhere==0.9.1
pytz==2019.3
requests==2.23.0
schema==0.7.1
six==1.14.0
sqlparse==0.3.1
tabulate==0.8.7
urllib3==1.25.8
wrapt==1.11.2
django-heroku==0.3.1